+++ src/App.jsx
@@ -6,17 +6,30 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Index from "./pages/Index";
 import NotFound from "./pages/NotFound";
 
+// Create a new QueryClient instance for React Query
 const queryClient = new QueryClient();
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
     <TooltipProvider>
+      {/* Toast notification providers */}
       <Toaster />
       <Sonner />
+      
+      {/* Application routing */}
       <BrowserRouter>
         <Routes>
+          {/* Main index route */}
           <Route path="/" element={<Index />} />
-          {/* ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL "*" ROUTE */}
+          
+          {/* 
+            Add all your custom routes above this catch-all route
+            Example:
+            <Route path="/dashboard" element={<Dashboard />} />
+            <Route path="/settings" element={<Settings />} />
+          */}
+          
+          {/* Catch-all route for 404 pages */}
           <Route path="*" element={<NotFound />} />
         </Routes>
       </BrowserRouter>
