+++ src/pages/Index.jsx
@@ -1,4 +1,3 @@
-
 import { useState } from 'react';
 import { PharmacyDashboard } from '@/components/pharmacy/PharmacyDashboard';
