+++ src/pages/NotFound.jsx
@@ -16,7 +16,10 @@ const NotFound = () => {
       <div className="text-center">
         <h1 className="text-4xl font-bold mb-4">404</h1>
         <p className="text-xl text-gray-600 mb-4">Oops! Page not found</p>
-        <a href="/" className="text-blue-500 hover:text-blue-700 underline">
+        <a 
+          href="/" 
+          className="text-blue-500 hover:text-blue-700 underline transition-colors"
+        >
           Return to Home
         </a>
       </div>
